==> Benchmarks/SCENIC+/0.1_export_cellTopic.ipynb <==
import pandas as pd
cistopic_obj = pd.read_pickle('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/pycisTopic/DPCL_cisTopicObject.pkl')

model = cistopic_obj.selected_model
cell_topic = model.cell_topic
cell_topic

import pandas as pd
cell_topic.to_csv('/data/leuven/312/vsc31245/jupyterhub_notebooks/SCENICplus/FigR_encode/analysis/cell_topic.txt', sep='\t')

umap = cistopic_obj.projections['cell']['UMAP']
umap.to_csv('/data/leuven/312/vsc31245/jupyterhub_notebooks/SCENICplus/FigR_encode/analysis/cell_umap.txt', sep='\t')

cistopic_obj.cell_data.to_csv('/data/leuven/312/vsc31245/jupyterhub_notebooks/SCENICplus/FigR_encode/analysis/cell_data.txt', sep='\t')
