==> Benchmarks/SCENIC+/3.SCENIC+-V10-V2.ipynb <==
### 1. Create SCENIC+ object

# Load functions
from scenicplus.scenicplus_class import SCENICPLUS, create_SCENICPLUS_object
from scenicplus.preprocessing.filtering import *

First we will load the scRNA-seq and the scATAC-seq data. We make sure that names match between them.

# Load data
## ATAC - cisTopic object
outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/pycisTopic/'
import pickle
infile = open(outDir + 'DPCL_cisTopicObject.pkl', 'rb')
cistopic_obj = pickle.load(infile)
infile.close()
## Precomputed imputed data
import pickle
infile = open(outDir + 'DARs/Imputed_accessibility.pkl', 'rb')
imputed_acc_obj = pickle.load(infile)
infile.close()
## RNA - Create Anndata
from loomxpy.loomxpy import SCopeLoom
from pycisTopic.loom import *
import itertools
import anndata
path_to_loom = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/vsn/grnboost/out/data/scRNA_count_matrix.SINGLE_SAMPLE_SCENIC.loom'
loom = SCopeLoom.read_loom(path_to_loom)
metadata = get_metadata(loom)
# Fix names
metadata.index =  metadata.index + '___DPLC'
expr_mat = loom.ex_mtx
expr_mat.index =  expr_mat.index + '___DPLC'
rna_anndata = anndata.AnnData(X=expr_mat)
rna_anndata.obs = metadata

Next we load the motif enrichment results into a dictionary. We can load motif results from the different methods in pycistarget (e.g. cisTarget, DEM) and different region sets (e.g. topics, DARs, MACS bdgdiff peaks). In this tutorial we will use both cisTarget and DEM peaks from topics and DARs.

## Precomputed imputed data
import pickle
infile = open('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/pycistarget/cluster_V10_V2/menr.pkl', 'rb') 
menr = pickle.load(infile)
infile.close()

Now we can create the SCENIC+ object:

scplus_obj = create_SCENICPLUS_object(
        GEX_anndata = rna_anndata,
        cisTopic_obj = cistopic_obj,
        imputed_acc_obj = imputed_acc_obj,
        menr = menr,
        ACC_prefix = 'ACC_',
        GEX_prefix = 'GEX_',
        bc_transform_func = lambda x: x,
        normalize_imputed_acc = False)

type(scplus_obj.X_EXP)

print(scplus_obj)

You can also filter low accessible regions and low expressed genes. This recommended to avoid getting false relationships with these regions and genes.

filter_genes(scplus_obj, min_pct = 0.5)
filter_regions(scplus_obj, min_pct = 0.5)

# Save
outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final/'
import pickle
with open(outDir+'scplus_obj.pkl', 'wb') as f:
  pickle.dump(scplus_obj, f)

### GRNBoost

singularity exec -B /lustre1,/staging,/data,/vsc-hard-mounts,/scratch scenicplus.sif ipython3

# For the downstream analyses
outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final/'
import pickle
infile = open(outDir+'scplus_obj.pkl', 'rb')
scplus_obj = pickle.load(infile)
infile.close()

import pickle
infile = open('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_V2_grnboost/region_ranking.pkl', 'rb')
region_ranking = pickle.load(infile)
infile.close()

import pickle
infile = open('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_V2_grnboost/gene_ranking.pkl', 'rb')
gene_ranking = pickle.load(infile)
infile.close()

from scenicplus.wrappers.run_scenicplus import *
run_scenicplus(scplus_obj,
    variable = ['ACC_Cell_type'],
    species = 'hsapiens',
    assembly = 'hg38',
    tf_file = '/staging/leuven/stg_00002/lcb/cflerin/resources/allTFs_hg38.txt',
    save_path = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final_autoreg/grnboost/',
    biomart_host = 'http://oct2016.archive.ensembl.org/',
    upstream = [1000, 150000],
    downstream = [1000, 150000],   
    calculate_TF_eGRN_correlation = False,
    calculate_DEGs_DARs = True,
    export_to_loom_file = True,
    export_to_UCSC_file = True,
    region_ranking=region_ranking,
    gene_ranking=gene_ranking,
    tree_structure = ('DPCL', 'SCENIC+', 'grnboost'),
    path_bedToBigBed = '/data/leuven/software/biomed/haswell_centos7/2018a/software/Kent_tools/20190730-linux.x86_64/bin/',
    n_cpu = 20,
    _temp_dir = '/scratch/leuven/313/vsc31305/ray_spill'
    )

# For the downstream analyses
outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final_autoreg/grnboost/'
import dill
infile = open(outDir+'scplus_obj.pkl', 'rb')
scplus_obj = dill.load(infile)
infile.close()

import pandas as pd
def format_egrns_time(scplus_obj,
                 eregulons_key: str = 'eRegulons',
                 TF2G_key: str = 'TF2G_adj',
                 key_added: str = 'eRegulon_metadata'):
    """
    A function to format eRegulons to a pandas dataframe
    """

    egrn_list = scplus_obj.uns[eregulons_key]
    TF = [egrn_list[x].transcription_factor for x in range(len(egrn_list))]
    is_extended = [str(egrn_list[x].is_extended)
                   for x in range(len(egrn_list))]
    r2g_data = [pd.DataFrame.from_records(egrn_list[x].regions2genes, columns=[
                                          'Region', 'Gene', 'R2G_importance', 'R2G_rho', 'R2G_importance_x_rho', 'R2G_importance_x_abs_rho']) for x in range(len(egrn_list))]
    egrn_name = [TF[x] + '_extended' if is_extended[x] ==
                 'True' else TF[x] for x in range(len(egrn_list))]
    egrn_name = [egrn_name[x] + '_+' if 'positive tf2g' in egrn_list[x]
                 .context else egrn_name[x] + '_-' for x in range(len(egrn_list))]
    egrn_name = [egrn_name[x] + '_+' if 'positive r2g' in egrn_list[x]
                 .context else egrn_name[x] + '_-' for x in range(len(egrn_list))]
    region_signature_name = [
        egrn_name[x] + '_(' + str(len(set(r2g_data[x].Region))) + 'r)' for x in range(len(egrn_list))]
    gene_signature_name = [
        egrn_name[x] + '_(' + str(len(set(r2g_data[x].Gene))) + 'g)' for x in range(len(egrn_list))]

    for x in range(len(egrn_list)):
        r2g_data[x].insert(0, "TF", TF[x])
        r2g_data[x].insert(1, "is_extended", is_extended[x])
        r2g_data[x].insert(0, "Gene_signature_name", gene_signature_name[x])
        r2g_data[x].insert(0, "Region_signature_name",
                           region_signature_name[x])

    tf2g_data = scplus_obj.uns[TF2G_key].copy()
    tf2g_data.columns = ['TF', 'Gene', 'TF2G_importance', 'TF2G_regulation',
                         'TF2G_rho', 'TF2G_importance_x_abs_rho', 'TF2G_importance_x_rho']
    egrn_metadata = pd.concat([pd.merge(r2g_data[x], tf2g_data[tf2g_data.TF == r2g_data[x].TF[0]], on=[
                              'TF', 'Gene']) for x in range(len(egrn_list)) if tf2g_data[tf2g_data.TF == r2g_data[x].TF[0]].shape[0] != 0 and r2g_data[x].shape[0] != 0])
    scplus_obj.uns[key_added] = egrn_metadata

format_egrns_time(scplus_obj,
              eregulons_key = 'eRegulons',
              TF2G_key = 'TF2G_adj',
              key_added = 'eRegulon_metadata')

scplus_obj.uns['region_to_gene'].to_csv('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_V2_grnboost_2/region_to_gene.tsv', sep='\t')



scplus_obj.uns['region_to_gene'][scplus_obj.uns['region_to_gene']['rho'] >0].to_csv('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_V2_grnboost_2/region_to_gene_pos.tsv', sep='\t')



select = list(set(scplus_obj.uns['eRegulon_metadata']['Region']))

scplus_obj.uns['region_to_gene']

scplus_obj.uns['region_to_gene'][scplus_obj.uns['region_to_gene']['region'].isin(select)]

scplus_obj.uns['region_to_gene'][scplus_obj.uns['region_to_gene']['region'].isin(select)].to_csv('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_V2_grnboost_2/region_to_gene_in_eGRN.tsv', sep='\t')

import pandas as pd
hic_data = pd.read_csv('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/formatted/HepG2_ENCFF020DPP_5Kb_SCALE.txt', sep='\t')
hic_data.columns = ['Chromosome', 'Start', 'End', 'target', 'importance', 'TSS', 'Strand', 'Distance', 'region']
hic_data['rho'] = hic_data['importance']
hic_data = hic_data[['target', 'region', 'importance', 'rho', 'Distance']]
hic_data = hic_data[hic_data['region'].isin(scplus_obj.region_names)]
hic_data 

import pandas as pd
import numpy as np
import ray
import logging
import time
import sys
import os
import subprocess
import pyranges as pr
from sklearn.ensemble import GradientBoostingRegressor, RandomForestRegressor, ExtraTreesRegressor
from scipy.stats import pearsonr, spearmanr
from tqdm import tqdm
from matplotlib import cm
from matplotlib.colors import Normalize
from typing import List

from scenicplus.utils import extend_pyranges, extend_pyranges_with_limits, reduce_pyranges_with_limits_b
from scenicplus.utils import calculate_distance_with_limits_join, reduce_pyranges_b, calculate_distance_join
from scenicplus.utils import coord_to_region_names, region_names_to_coordinates, ASM_SYNONYMS, Groupby, flatten_list
from scenicplus.scenicplus_class import SCENICPLUS
from scenicplus.enhancer_to_gene import INTERACT_AS



def export_to_UCSC_interact_hic(SCENICPLUS_obj: SCENICPLUS,
                            species: str,
                            outfile: str,
                            region_to_gene_key: str =' region_to_gene',
                            pbm_host:str = 'http://www.ensembl.org',
                            bigbed_outfile:str = None,
                            path_bedToBigBed: str= None,
                            assembly: str = None,
                            ucsc_track_name: str = 'region_to_gene',
                            ucsc_description: str = 'interaction file for region to gene',
                            cmap_neg: str = 'Reds',
                            cmap_pos: str = 'Greens',
                            key_for_color: str = 'importance',
                            vmin: int = 0,
                            vmax: int = 1,
                            scale_by_gene: bool = True,
                            subset_for_eRegulons_regions: bool = True,
                            eRegulons_key: str = 'eRegulons') -> pd.DataFrame:
    # Create logger
    level = logging.INFO
    format = '%(asctime)s %(name)-12s %(levelname)-8s %(message)s'
    handlers = [logging.StreamHandler(stream=sys.stdout)]
    logging.basicConfig(level=level, format=format, handlers=handlers)
    log = logging.getLogger('R2G')

    if region_to_gene_key not in SCENICPLUS_obj.uns.keys():
        raise Exception(
            f'key {region_to_gene_key} not found in SCENICPLUS_obj.uns, first calculate region to gene relationships using function: "calculate_regions_to_genes_relationships"')

    region_to_gene_df = SCENICPLUS_obj.uns[region_to_gene_key].copy()

    if subset_for_eRegulons_regions:
        if eRegulons_key not in SCENICPLUS_obj.uns.keys():
            raise ValueError(
                f'key {eRegulons_key} not found in SCENICPLUS_obj.uns.keys()')
        eRegulon_regions = list(set(flatten_list(
            [ereg.target_regions for ereg in SCENICPLUS_obj.uns[eRegulons_key]])))
        region_to_gene_df.index = region_to_gene_df['region']
        region_to_gene_df = region_to_gene_df.loc[eRegulon_regions].reset_index(
            drop=True)

    # Rename columns to be in line with biomart annotation
    region_to_gene_df.rename(columns={'target': 'Gene'}, inplace=True)

    # Get TSS annotation (end-point for links)
    log.info('Downloading gene annotation from biomart, using dataset: {}'.format(
        species+'_gene_ensembl'))
    import pybiomart as pbm
    dataset = pbm.Dataset(name=species+'_gene_ensembl',  host=pbm_host)
    annot = dataset.query(attributes=['chromosome_name', 'start_position', 'end_position',
                          'strand', 'external_gene_name', 'transcription_start_site', 'transcript_biotype'])
    annot.columns = ['Chromosome', 'Start', 'End', 'Strand',
                     'Gene', 'Transcription_Start_Site', 'Transcript_type']
    annot['Chromosome'] = 'chr' + \
        annot['Chromosome'].astype(str)
    annot = annot[annot.Transcript_type == 'protein_coding']
    annot.Strand[annot.Strand == 1] = '+'
    annot.Strand[annot.Strand == -1] = '-'

    log.info('Formatting data ...')
    # get gene to tss mapping, take the one equal to the gene start/end location if possible otherwise take the first one
    annot['TSSeqStartEnd'] = np.logical_or(
        annot['Transcription_Start_Site'] == annot['Start'], annot['Transcription_Start_Site'] == annot['End'])
    gene_to_tss = annot[['Gene', 'Transcription_Start_Site']].groupby(
        'Gene').agg(lambda x: list(map(str, x)))
    startEndEq = annot[['Gene', 'TSSeqStartEnd']
                       ].groupby('Gene').agg(lambda x: list(x))
    gene_to_tss['Transcription_Start_Site'] = [np.array(tss[0])[eq[0]][0] if sum(
        eq[0]) >= 1 else tss[0][0] for eq, tss in zip(startEndEq.values, gene_to_tss.values)]
    gene_to_tss.columns = ['TSS_Gene']

    # get gene to strand mapping
    gene_to_strand = annot[['Gene', 'Strand']].groupby(
        'Gene').agg(lambda x: list(map(str, x))[0])

    # get gene to chromosome mapping (should be the same as the regions mapped to the gene)
    gene_to_chrom = annot[['Gene', 'Chromosome']].groupby(
        'Gene').agg(lambda x: list(map(str, x))[0])

    # add TSS for each gene to region_to_gene_df
    region_to_gene_df = region_to_gene_df.join(gene_to_tss, on='Gene')

    # add strand for each gene to region_to_gene_df
    region_to_gene_df = region_to_gene_df.join(gene_to_strand, on='Gene')

    # add chromosome for each gene to region_to_gene_df
    region_to_gene_df = region_to_gene_df.join(gene_to_chrom, on='Gene')

    # get chrom, chromStart, chromEnd
    region_to_gene_df.dropna(axis=0, how='any', inplace=True)
    arr = region_names_to_coordinates(region_to_gene_df['region']).to_numpy()
    chrom, chromStart, chromEnd = np.split(arr, 3, 1)
    chrom = chrom[:, 0]
    chromStart = chromStart[:, 0]
    chromEnd = chromEnd[:, 0]

    # get source chrom, chromStart, chromEnd (i.e. middle of regions)
    sourceChrom = chrom
    sourceStart = np.array(
        list(map(int, chromStart + (chromEnd - chromStart)/2 - 1)))
    sourceEnd = np.array(
        list(map(int, chromStart + (chromEnd - chromStart)/2)))

    # get target chrom, chromStart, chromEnd (i.e. TSS)
    targetChrom = region_to_gene_df['Chromosome']
    targetStart = region_to_gene_df['TSS_Gene'].values
    targetEnd = list(map(str, np.array(list(map(int, targetStart))) + np.array(
        [1 if strand == '+' else -1 for strand in region_to_gene_df['Strand'].values])))

    # get color
    norm = Normalize(vmin=vmin, vmax=vmax)
    if scale_by_gene:
        grouper = Groupby(
            region_to_gene_df.loc[:, 'Gene'].to_numpy())
        scores = region_to_gene_df.loc[:, key_for_color].to_numpy()
        mapper = cm.ScalarMappable(norm=norm, cmap=cmap_pos)

        def _value_to_color(scores):
            S = (scores - scores.min()) / (scores.max() - scores.min())
            return [','.join([str(x) for x in mapper.to_rgba(s, bytes=True)][0:3]) for s in S]

        colors_pos = np.zeros(len(scores), dtype='object')
        for idx in grouper.indices:
            colors_pos[idx] = _value_to_color(scores[idx])

        def _value_to_color(scores):
            S = (scores - scores.min()) / (scores.max() - scores.min())
            return [','.join([str(x) for x in mapper.to_rgba(s, bytes=True)][0:3]) for s in S]

        colors_neg = np.zeros(len(scores), dtype='object')
        for idx in grouper.indices:
            colors_neg[idx] = _value_to_color(scores[idx])

    else:
        scores = region_to_gene_df.loc[:, key_for_color].to_numpy()
        mapper = cm.ScalarMappable(norm=norm, cmap=cmap_pos)
        colors_pos = [
            ','.join([str(x) for x in mapper.to_rgba(s, bytes=True)][0:3]) for s in scores]


    region_to_gene_df.loc[:, 'color'] = colors_pos
    region_to_gene_df['color'] = region_to_gene_df['color'].fillna('55,55,55')

    # get name for regions (add incremental number to gene in range of regions linked to gene)
    counter = 1
    previous_gene = region_to_gene_df['Gene'].values[0]
    names = []
    for gene in region_to_gene_df['Gene'].values:
        if gene != previous_gene:
            counter = 1
        else:
            counter += 1
        names.append(gene + '_' + str(counter))
        previous_gene = gene

    # format final interact dataframe
    df_interact = pd.DataFrame(
        data={
            'chrom':        chrom,
            'chromStart':   chromStart,
            'chromEnd':     chromEnd,
            'name':         names,
            'score':        (1000*(region_to_gene_df['importance'].values - np.min(region_to_gene_df['importance'].values))/np.ptp(region_to_gene_df['importance'].values)).astype(int) ,
            'value':        region_to_gene_df['importance'].values,
            'exp':          np.repeat('.', len(region_to_gene_df)),
            'color':        region_to_gene_df['color'].values,
            'sourceChrom':  sourceChrom,
            'sourceStart':  sourceStart,
            'sourceEnd':    sourceEnd,
            'sourceName':   names,
            'sourceStrand': np.repeat('.', len(region_to_gene_df)),
            'targetChrom':  targetChrom,
            'targetStart':  targetStart,
            'targetEnd':    targetEnd,
            'targetName':   region_to_gene_df['Gene'].values,
            'targetStrand': region_to_gene_df['Strand'].values
        }
    )
    # sort dataframe
    df_interact = df_interact.sort_values(by=['chrom', 'chromStart'])
    # Write interact file
    log.info('Writing data to: {}'.format(outfile))
    with open(outfile, 'w') as f:
        f.write('track type=interact name="{}" description="{}" useScore=0 maxHeightPixels=200:100:50 visibility=full\n'.format(
            ucsc_track_name, ucsc_description))
        df_interact.to_csv(f, header=False, index=False, sep='\t')

    # write bigInteract file
    if bigbed_outfile != None:
        log.info('Writing data to: {}'.format(bigbed_outfile))
        outfolder = bigbed_outfile.rsplit('/', 1)[0]
        # write bed file without header to tmp file
        df_interact.to_csv(os.path.join(
            outfolder, 'interact.bed.tmp'), header=False, index=False, sep='\t')

        # check if auto sql definition for interaction file exists in outfolder, otherwise create it
        if not os.path.exists(os.path.join(outfolder, 'interact.as')):
            with open(os.path.join(outfolder, 'interact.as'), 'w') as f:
                f.write(INTERACT_AS)
        # convert interact.bed.tmp to bigBed format
        # bedToBigBed -as=interact.as -type=bed5+13 region_to_gene_no_head.interact https://genome.ucsc.edu/goldenPath/help/hg38.chrom.sizes region_to_gene.inter.bb
        cmds = [
            os.path.join(path_bedToBigBed, 'bedToBigBed'),
            '-as={}'.format(os.path.join(os.path.join(outfolder, 'interact.as'))),
            '-type=bed5+13',
            os.path.join(outfolder, 'interact.bed.tmp'),
            'https://hgdownload.cse.ucsc.edu/goldenpath/' + assembly + '/bigZips/' + assembly + '.chrom.sizes',
            bigbed_outfile
        ]
        p = subprocess.Popen(cmds, stdout=subprocess.PIPE,
                             stderr=subprocess.PIPE)
        stdout, stderr = p.communicate()
        if p.returncode:
            raise ValueError(
                "cmds: %s\nstderr:%s\nstdout:%s" % (
                    " ".join(cmds), stderr, stdout)
            )
    return df_interact

# Generate Hi-C files

# For the downstream analyses
outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final_autoreg/grnboost/'
import dill
infile = open(outDir+'scplus_obj.pkl', 'rb')
scplus_obj = dill.load(infile)
infile.close()

r2g = scplus_obj.uns['region_to_gene'].copy()
r2g = r2g[r2g.rho > 0.03]
r2g

scplus_obj.uns['region_to_gene'].to_csv('/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final_autoreg/grnboost/region_to_gene_pos.tsv', sep='\t')

import os
from os import listdir
from os.path import isfile, join
import pandas as pd
cell_lines=['GM12878', 'HepG2', 'IMR90', 'HCT116', 'K562']
save_path = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/bb_files_final/'
species = 'hsapiens'
assembly = 'hg38'
path_bedToBigBed = '/data/leuven/software/biomed/haswell_centos7/2018a/software/Kent_tools/20190730-linux.x86_64/bin/'
biomart_host = 'http://oct2016.archive.ensembl.org/'
path='/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/formatted/'
for cell_line in cell_lines:
    files = [f for f in listdir(path) if isfile(join(path, f))]
    file = [f for f in files if cell_line in f]
    hic_data = pd.read_csv(path+file[0], sep='\t')
    hic_data.columns = ['Chromosome', 'Start', 'End', 'target', 'importance', 'TSS', 'Strand', 'Distance', 'region']
    hic_data['rho'] = hic_data['importance']
    hic_data = hic_data[['target', 'region', 'importance', 'rho', 'Distance']]
    hic_data = hic_data[hic_data['region'].isin(scplus_obj.region_names)]
    scplus_obj.uns['region_to_gene'] = hic_data
    r2g_data = export_to_UCSC_interact_hic(scplus_obj,
                    species,
                    os.path.join(save_path,cell_line+'.hic.all.bed'),
                    path_bedToBigBed=path_bedToBigBed,
                    bigbed_outfile=os.path.join(save_path,cell_line+'.hic.all.bb'),
                    region_to_gene_key='region_to_gene',
                    pbm_host=biomart_host,
                    assembly=assembly,
                    ucsc_track_name='R2G',
                    ucsc_description=cell_line+' HiC links',
                    cmap_neg='Reds',
                    cmap_pos='Greys',
                    key_for_color='importance',
                    scale_by_gene=True,
                    subset_for_eRegulons_regions=False,
                    eRegulons_key='eRegulons')

import os
from os import listdir
from os.path import isfile, join
import pandas as pd
cell_lines=['CellOracle', 'FigR', 'GRaNIE', 'Scenicplus-importance', 'Scenicplus-rho']
color_dict={'CellOracle':'Greens', 'FigR':'Purples', 'GRaNIE':'Oranges', 'Scenicplus-importance':'Blues', 'Scenicplus-rho':'Reds'}
save_path = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/bb_files_final/'
species = 'hsapiens'
assembly = 'hg38'
path_bedToBigBed = '/data/leuven/software/biomed/haswell_centos7/2018a/software/Kent_tools/20190730-linux.x86_64/bin/'
biomart_host = 'http://oct2016.archive.ensembl.org/'
path='/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/formatted/'
for cell_line in cell_lines:
    files = [f for f in listdir(path) if isfile(join(path, f))]
    file = [f for f in files if cell_line in f]
    hic_data = pd.read_csv(path+file[0], sep='\t')
    hic_data.columns = ['Chromosome', 'Start', 'End', 'target', 'importance', 'TSS', 'Strand', 'Distance', 'region']
    hic_data['rho'] = hic_data['importance']
    hic_data = hic_data[['target', 'region', 'importance', 'rho', 'Distance']]
    hic_data = hic_data[hic_data['region'].isin(scplus_obj.region_names)]
    scplus_obj.uns['region_to_gene'] = hic_data
    r2g_data = export_to_UCSC_interact_hic(scplus_obj,
                    species,
                    os.path.join(save_path,cell_line+'.links.all.bed'),
                    path_bedToBigBed=path_bedToBigBed,
                    bigbed_outfile=os.path.join(save_path,cell_line+'.links.all.bb'),
                    region_to_gene_key='region_to_gene',
                    pbm_host=biomart_host,
                    assembly=assembly,
                    ucsc_track_name='R2G',
                    ucsc_description='SCENIC+ region to gene links',
                    cmap_neg='Reds',
                    cmap_pos=color_dict[cell_line],
                    key_for_color='importance',
                    scale_by_gene=True,
                    subset_for_eRegulons_regions=False,
                    eRegulons_key='eRegulons')


import os
from os import listdir
from os.path import isfile, join
import pandas as pd
cell_lines=['Scenicplus-importance_links_all', 'Scenicplus-rho_links_all']
color_dict={'Scenicplus-importance_links_all':'Blues', 'Scenicplus-rho_links_all':'Reds'}
save_path = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/bb_files_final/'
species = 'hsapiens'
assembly = 'hg38'
path_bedToBigBed = '/data/leuven/software/biomed/haswell_centos7/2018a/software/Kent_tools/20190730-linux.x86_64/bin/'
biomart_host = 'http://oct2016.archive.ensembl.org/'
path='/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/data/HiC/formatted/'
for cell_line in cell_lines:
    files = [f for f in listdir(path) if isfile(join(path, f))]
    file = [f for f in files if cell_line in f]
    hic_data = pd.read_csv(path+file[0], sep='\t')
    hic_data.columns = ['Chromosome', 'Start', 'End', 'target', 'importance', 'TSS', 'Strand', 'Distance', 'region']
    hic_data['rho'] = hic_data['importance']
    hic_data = hic_data[['target', 'region', 'importance', 'rho', 'Distance']]
    hic_data = hic_data[hic_data['region'].isin(scplus_obj.region_names)]
    scplus_obj.uns['region_to_gene'] = hic_data
    r2g_data = export_to_UCSC_interact_hic(scplus_obj,
                    species,
                    os.path.join(save_path,cell_line+'.links.all.bed'),
                    path_bedToBigBed=path_bedToBigBed,
                    bigbed_outfile=os.path.join(save_path,cell_line+'.links.all.bb'),
                    region_to_gene_key='region_to_gene',
                    pbm_host=biomart_host,
                    assembly=assembly,
                    ucsc_track_name='R2G',
                    ucsc_description='SCENIC+ region to gene links',
                    cmap_neg='Reds',
                    cmap_pos=color_dict[cell_line],
                    key_for_color='importance',
                    scale_by_gene=True,
                    subset_for_eRegulons_regions=False,
                    eRegulons_key='eRegulons')





#from scenicplus.enhancer_to_gene import export_to_UCSC_interact 


hic_data 

eRegulon_regions = list(set(flatten_list([ereg.target_regions for ereg in scplus_obj.uns['eRegulons']])))
hic_data = hic_data[hic_data['region'].isin(eRegulon_regions)]
scplus_obj.uns['region_to_gene'] = hic_data
r2g_data = export_to_UCSC_interact_hic(scplus_obj,
                    species,
                    os.path.join(save_path,'HepG2.hic.eGRN.bed'),
                    path_bedToBigBed=path_bedToBigBed,
                    bigbed_outfile=os.path.join(save_path,'HepG2.hic.eGRN.notscaled.bb'),
                    region_to_gene_key='region_to_gene',
                    pbm_host=biomart_host,
                    assembly=assembly,
                    ucsc_track_name='R2G',
                    ucsc_description='SCENIC+ region to gene links',
                    cmap_neg='Reds',
                    cmap_pos='Greens',
                    key_for_color='importance',
                    scale_by_gene=False,
                    subset_for_eRegulons_regions=False,
                    eRegulons_key='eRegulons')

# For the downstream analyses

#outDir = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final/'
outDir = '/data/users/cbravo/DPCL/scenicplus_final_autoreg/'
import pickle
infile = open(outDir+'scplus_obj.pkl', 'rb')
scplus_obj = pickle.load(infile)
infile.close()

import pickle
infile = open('/data/users/cbravo/DPCL/scenicplus_final_autoreg/genie3/region_ranking.pkl', 'rb')
region_ranking = pickle.load(infile)
infile.close()

import pickle
infile = open('/data/users/cbravo/DPCL/scenicplus_final_autoreg/genie3/gene_ranking.pkl', 'rb')
gene_ranking = pickle.load(infile)
infile.close()

from scenicplus.wrappers.run_scenicplus import *
run_scenicplus_genie3(scplus_obj,
    variable = ['ACC_Cell_type'],
    species = 'hsapiens',
    assembly = 'hg38',
    tf_file = '/data/users/cbravo/resources/allTFs_hg38.txt',
    save_path = '/data/users/cbravo/DPCL/scenicplus_final_autoreg/genie3/',
    biomart_host = 'http://oct2016.archive.ensembl.org/',
    upstream = [1000, 150000],
    downstream = [1000, 150000],   
    calculate_TF_eGRN_correlation = False,
    calculate_DEGs_DARs = True,
    export_to_loom_file = True,
    export_to_UCSC_file = True,
    region_ranking=region_ranking,
    gene_ranking=gene_ranking,
    tree_structure = ('DPCL', 'SCENIC+', 'genie3'),
    path_bedToBigBed = '/media/data/users/cbravo/software/KENT/',
    n_cpu = 20,
    _temp_dir = '/media/data/users/cbravo/ray_spill'
    )

from scenicplus.wrappers.run_scenicplus import *
run_scenicplus_genie3(scplus_obj,
    variable = ['ACC_Cell_type'],
    species = 'hsapiens',
    assembly = 'hg38',
    tf_file = '/staging/leuven/stg_00002/lcb/cflerin/resources/allTFs_hg38.txt',
    save_path = '/staging/leuven/stg_00002/lcb/cbravo/Multiomics_pipeline/analysis/DPCL/scenicplus_final_autoreg/genie3/',
    biomart_host = 'http://oct2016.archive.ensembl.org/',
    upstream = [1000, 150000],
    downstream = [1000, 150000],   
    calculate_TF_eGRN_correlation = False,
    calculate_DEGs_DARs = True,
    export_to_loom_file = True,
    export_to_UCSC_file = True,
    tree_structure = ('DPCL', 'SCENIC+', 'genie3'),
    path_bedToBigBed = '/data/leuven/software/biomed/haswell_centos7/2018a/software/Kent_tools/20190730-linux.x86_64/bin/',
    n_cpu = 14,
    _temp_dir = '/scratch/leuven/313/vsc31305/ray_spill'
    )

from scenicplus.scenicplus_class import SCENICPLUS, create_SCENICPLUS_object
from scenicplus.preprocessing.filtering import *
from scenicplus.cistromes import *
from scenicplus.enhancer_to_gene import get_search_space, calculate_regions_to_genes_relationships, RF_KWARGS
from scenicplus.enhancer_to_gene import export_to_UCSC_interact 
from scenicplus.utils import format_egrns, export_eRegulons
from scenicplus.eregulon_enrichment import *
from scenicplus.TF_to_gene import *
from scenicplus.grn_builder.gsea_approach import build_grn
from scenicplus.dimensionality_reduction import *
from scenicplus.RSS import *
from scenicplus.diff_features import *
from scenicplus.loom import *
from typing import Dict, List, Mapping, Optional, Sequence
import os
import dill
import time

def run_scenicplus_genie3(scplus_obj: 'SCENICPLUS',
    variable: List[str],
    species: str,
    assembly: str,
    tf_file: str,
    save_path: str,
    biomart_host: Optional[str] = 'http://www.ensembl.org',
    upstream: Optional[List] = [1000, 150000],
    downstream: Optional[List] = [1000, 150000],
    region_ranking: Optional['CisTopicImputedFeatures'] = None,
    gene_ranking: Optional['CisTopicImputedFeatures'] = None,
    calculate_TF_eGRN_correlation: Optional[bool] = True,
    calculate_DEGs_DARs: Optional[bool] = True,
    export_to_loom_file: Optional[bool] = True,
    export_to_UCSC_file: Optional[bool] = True,
    tree_structure: Sequence[str] = (),
    path_bedToBigBed: Optional[str] = None,
    n_cpu: Optional[int] = 1,
    _temp_dir: Optional[str] = '/scratch/leuven/313/vsc31305/ray_spill'
    ):
    """
    Wrapper to run SCENIC+
    
    Parameters
    ---------
    scplus_obj: `class::SCENICPLUS`
        A SCENICPLUS object.
    variables: List[str]
        Variables to use for RSS, TF-eGRN correlation and markers.
    species: str
        Species from which data comes from. Possible values: 'hsapiens', 'mmusculus', 'dmelanogaster'
    assembly: str
        Genome assembly to which the data was mapped. Possible values: 'hg38'
    tf_file: str
        Path to file containing genes that are TFs
    save_path: str
        Folder in which results will be saved
    biomart_host: str, optional
        Path to biomart host. Make sure that the host matches your genome assembly
    upstream: str, optional
        Upstream space to use for region to gene relationships
    downstream: str, optional
        Upstream space to use for region to gene relationships
    region_ranking: `class::CisTopicImputedFeatures`, optional
        Precomputed region ranking
    gene_ranking: `class::CisTopicImputedFeatures`, optional
        Precomputed gene ranking
    calculate_TF_eGRN_correlation: bool, optional
        Whether to calculate the TF-eGRN correlation based on the variables
    calculate_DEGs_DARs: bool, optional
        Whether to calculate DARs/DEGs based on the variables
    export_to_loom_file: bool, optional
        Whether to export data to loom files (gene based/region based)
    export_to_UCSC_file: bool, optional
        Whether to export region-to-gene links and eregulons to bed files
    tree_structure: sequence, optional
        Tree structure for loom files
    path_bedToBigBed: str, optional
        Path to convert bed files to big bed when exporting to UCSC (required if files are meant to be
        used in a hub)
    n_cpu: int, optional
        Number of cores to use
    _temp_dir: str, optional
        Temporary directory for ray
    """
    
    # Create logger
    level = logging.INFO
    log_format = '%(asctime)s %(name)-12s %(levelname)-8s %(message)s'
    handlers = [logging.StreamHandler(stream=sys.stdout)]
    logging.basicConfig(level=level, format=log_format, handlers=handlers)
    log = logging.getLogger('SCENIC+_wrapper')
    
    start_time = time.time()
    
    check_folder = os.path.isdir(save_path)
    if not check_folder:
        os.makedirs(save_path)
        log.info("Created folder : "+ save_path)

    else:
        log.info(save_path + " folder already exists.")
    
    if 'Cistromes' not in scplus_obj.uns.keys():
        log.info('Merging cistromes')
        merge_cistromes(scplus_obj)
    
    
    if 'search_space' not in scplus_obj.uns.keys():
        log.info('Getting search space')
        get_search_space(scplus_obj,
                     biomart_host = biomart_host,
                     species = species,
                     assembly = assembly, 
                     upstream = upstream,
                     downstream = downstream)
                 
    if 'region_to_gene' not in scplus_obj.uns.keys():
        log.info('Inferring region to gene relationships')
        calculate_regions_to_genes_relationships(scplus_obj, 
                        ray_n_cpu = n_cpu, 
                        _temp_dir = _temp_dir,
                        importance_scoring_method = 'RF',
                        importance_scoring_kwargs = RF_KWARGS)
                        
    if 'TF2G_adj' not in scplus_obj.uns.keys():
        log.info('Inferring TF to gene relationships')
        calculate_TFs_to_genes_relationships(scplus_obj, 
                        tf_file = tf_file,
                        ray_n_cpu = n_cpu, 
                        method = 'GBM',
                        _temp_dir = _temp_dir,
                        key= 'TF2G_adj')
                        
    if 'eRegulons' not in scplus_obj.uns.keys():
        log.info('Build eGRN')
        build_grn(scplus_obj,
                 min_target_genes = 10,
                 adj_pval_thr = 1,
                 min_regions_per_gene = 0,
                 quantiles = (0.85, 0.90, 0.95),
                 top_n_regionTogenes_per_gene = (5, 10, 15),
                 top_n_regionTogenes_per_region = (),
                 binarize_using_basc = True,
                 rho_dichotomize_tf2g = True,
                 rho_dichotomize_r2g = True,
                 rho_dichotomize_eregulon = True,
                 rho_threshold = 0.05,
                 keep_extended_motif_annot = True,
                 merge_eRegulons = True, 
                 order_regions_to_genes_by = 'importance',
                 order_TFs_to_genes_by = 'importance',
                 key_added = 'eRegulons',
                 cistromes_key = 'Unfiltered',
                 disable_tqdm = True, 
                 ray_n_cpu = n_cpu,
                 _temp_dir = _temp_dir)

    if 'eRegulon_metadata' not in scplus_obj.uns.keys():
        log.info('Formatting eGRNs')
        format_egrns(scplus_obj,
                      eregulons_key = 'eRegulons',
                      TF2G_key = 'TF2G_adj',
                      key_added = 'eRegulon_metadata')


    if 'eRegulon_signatures' not in scplus_obj.uns.keys():
        log.info('Converting eGRNs to signatures')
        get_eRegulons_as_signatures(scplus_obj,
                                     eRegulon_metadata_key='eRegulon_metadata', 
                                     key_added='eRegulon_signatures')

#if 'eRegulon_AUC' not in scplus_obj.uns.keys():
    log.info('Calculating eGRNs AUC')
    if region_ranking is None:
        log.info('Calculating region ranking')
        region_ranking = make_rankings(scplus_obj, target='region')
        with open(os.path.join(save_path,'region_ranking.pkl'), 'wb') as f:
            dill.dump(region_ranking, f)
    log.info('Calculating eGRNs region based AUC')
    score_eRegulons(scplus_obj,
            ranking = region_ranking,
            eRegulon_signatures_key = 'eRegulon_signatures',
            key_added = 'eRegulon_AUC', 
            enrichment_type= 'region',
            auc_threshold = 0.05,
            normalize = False,
            n_cpu = n_cpu)
    if gene_ranking is None:
        log.info('Calculating gene ranking')
        gene_ranking = make_rankings(scplus_obj, target='gene')
        with open(os.path.join(save_path,'gene_ranking.pkl'), 'wb') as f:
            dill.dump(gene_ranking, f)
    log.info('Calculating eGRNs gene based AUC')
    score_eRegulons(scplus_obj,
            gene_ranking,
            eRegulon_signatures_key = 'eRegulon_signatures',
            key_added = 'eRegulon_AUC', 
            enrichment_type = 'gene',
            auc_threshold = 0.05,
            normalize= False,
            n_cpu = n_cpu)
                
    if calculate_TF_eGRN_correlation is True:
        log.info('Calculating TF-eGRNs AUC correlation')
        for var in variable:
            generate_pseudobulks(scplus_obj, 
                             variable = var,
                             auc_key = 'eRegulon_AUC',
                             signature_key = 'Gene_based',
                             nr_cells = 5,
                             nr_pseudobulks = 100,
                             seed=555)
            generate_pseudobulks(scplus_obj, 
                                     variable = var,
                                     auc_key = 'eRegulon_AUC',
                                     signature_key = 'Region_based',
                                     nr_cells = 5,
                                     nr_pseudobulks = 100,
                                     seed=555)
            TF_cistrome_correlation(scplus_obj,
                            variable = var, 
                            auc_key = 'eRegulon_AUC',
                            signature_key = 'Gene_based',
                            out_key = var+'_eGRN_gene_based')
            TF_cistrome_correlation(scplus_obj,
                                    variable = var, 
                                    auc_key = 'eRegulon_AUC',
                                    signature_key = 'Region_based',
                                    out_key = var+'_eGRN_region_based')
                                
#if 'eRegulon_AUC_thresholds' not in scplus_obj.uns.keys():
    log.info('Binarizing eGRNs AUC')
    binarize_AUC(scplus_obj, 
         auc_key='eRegulon_AUC',
         out_key='eRegulon_AUC_thresholds',
         signature_keys=['Gene_based', 'Region_based'],
         n_cpu=n_cpu)

#if 'eRegulons_UMAP' not in scplus_obj.dr_cell.keys():
    log.info('Making eGRNs AUC UMAP')
    run_eRegulons_umap(scplus_obj,
               scale=True, signature_keys=['Gene_based', 'Region_based'])
#if 'eRegulons_tSNE' not in scplus_obj.dr_cell.keys():
    log.info('Making eGRNs AUC tSNE')
    run_eRegulons_tsne(scplus_obj,
               scale=True, signature_keys=['Gene_based', 'Region_based'])

#if 'RSS' not in scplus_obj.uns.keys():
    log.info('Calculating eRSS')
    for var in variable:
        regulon_specificity_scores(scplus_obj, 
                     var,
                     signature_keys=['Gene_based'],
                     out_key_suffix='_gene_based',
                     scale=False)
        regulon_specificity_scores(scplus_obj, 
                     var,
                     signature_keys=['Region_based'],
                     out_key_suffix='_region_based',
                         scale=False)
                         
    if calculate_DEGs_DARs is True:
        log.info('Calculating DEGs/DARs')
        for var in variable:
            get_differential_features(scplus_obj, var, use_hvg = True, contrast_type = ['DEGs', 'DARs'])
            
    if export_to_loom_file is True:
        log.info('Exporting to loom file')
        export_to_loom(scplus_obj, 
               signature_key = 'Gene_based',
               tree_structure = tree_structure,
               title =  'Gene based eGRN',
               nomenclature = assembly,
               out_fname=os.path.join(save_path,'SCENIC+_gene_based.loom'))
        export_to_loom(scplus_obj, 
               signature_key = 'Region_based',
               tree_structure = tree_structure,
               title =  'Region based eGRN',
               nomenclature = assembly,
               out_fname=os.path.join(save_path,'SCENIC+_region_based.loom'))
               
    if export_to_UCSC_file is True:
        log.info('Exporting to UCSC')
        r2g_data = export_to_UCSC_interact(scplus_obj,
                            species,
                            os.path.join(save_path,'r2g.rho.bed'),
                            path_bedToBigBed=path_bedToBigBed,
                            bigbed_outfile=os.path.join(save_path,'r2g.rho.bb'),
                            region_to_gene_key='region_to_gene',
                            pbm_host=biomart_host,
                            assembly=assembly,
                            ucsc_track_name='R2G',
                            ucsc_description='SCENIC+ region to gene links',
                            cmap_neg='Reds',
                            cmap_pos='Greens',
                            key_for_color='rho',
                            scale_by_gene=False,
                            subset_for_eRegulons_regions=True,
                            eRegulons_key='eRegulons')
        r2g_data = export_to_UCSC_interact(scplus_obj,
                            species,
                            os.path.join(save_path,'r2g.importance.bed'),
                            path_bedToBigBed=path_bedToBigBed,
                            bigbed_outfile=os.path.join(save_path,'r2g.importance.bb'),
                            region_to_gene_key='region_to_gene',
                            pbm_host=biomart_host,
                            assembly=assembly,
                            ucsc_track_name='R2G',
                            ucsc_description='SCENIC+ region to gene links',
                            cmap_neg='Reds',
                            cmap_pos='Greens',
                            key_for_color='importance',
                            scale_by_gene=True,
                            subset_for_eRegulons_regions=True,
                            eRegulons_key='eRegulons')
        regions = export_eRegulons(scplus_obj,
                os.path.join(save_path,'eRegulons.bed'),
                assembly,
                bigbed_outfile = os.path.join(save_path,'eRegulons.bb'),
                eRegulon_metadata_key = 'eRegulon_metadata',
                eRegulon_signature_key = 'eRegulon_signatures',
                path_bedToBigBed=path_bedToBigBed)
        
    log.info('Saving object')         
    with open(os.path.join(save_path,'scplus_obj_genie3.pkl'), 'wb') as f:
        dill.dump(scplus_obj, f)

    log.info('Finished! Took {} minutes'.format((time.time() - start_time)/60))
